+++ tests/unit/data_types.test.ts
@@ -176,14 +176,14 @@ describe('Data Types Tests', () => {
 
   it('should correctly create a tuple data type with the correct typeStr', () => {
     const tuple = ClickhouseTypes.CHTuple(
-      ClickhouseTypes.CHFloat64(), ClickhouseTypes.CHFloat64()
+      [ClickhouseTypes.CHFloat64(), ClickhouseTypes.CHFloat64()]
     )
   
     expect(tuple.toString()).toEqual('Tuple(Float64, Float64)')
   })
   
   it('should correctly create a tuple data type with default value', () => {
-    const tuple = ClickhouseTypes.CHTuple(ClickhouseTypes.CHFloat64(), ClickhouseTypes.CHFloat64(), [10.5, 20.3])
+    const tuple = ClickhouseTypes.CHTuple([ClickhouseTypes.CHFloat64(), ClickhouseTypes.CHString(), ClickhouseTypes.CHBoolean()], [10.5, '', false])
   
     expect(tuple.toString()).toEqual('Tuple(Float64, Float64)')
     expect(tuple.getDefaultSql()).toEqual('(10.5, 20.3)')